@@ -11,49 +11,52 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.7",
-    "@angular/cdk": "^5.2.3",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "^5.0.0-beta.13",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.3",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "@aspnet/signalr": "^1.1.0",
+    "@angular/animations": "7.2.4",
+    "@angular/cdk": "7.3.2",
+    "@angular/common": "7.2.4",
+    "@angular/compiler": "7.2.4",
+    "@angular/core": "7.2.4",
+    "@angular/flex-layout": "^7.0.0-beta.23",
+    "@angular/forms": "7.2.4",
+    "@angular/http": "7.2.4",
+    "@angular/material": "7.3.2",
+    "@angular/platform-browser": "7.2.4",
+    "@angular/platform-browser-dynamic": "7.2.4",
+    "@angular/router": "7.2.4",
+    "@aspnet/signalr": "1.1.0",
     "@fortawesome/angular-fontawesome": "^0.3.0",
     "@fortawesome/fontawesome-svg-core": "^1.2.15",
     "@fortawesome/free-solid-svg-icons": "^5.7.2",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "@ngrx/store-devtools": "^5.2.0",
-    "@ngx-progressbar/core": "^4.1.1",
-    "@ngx-progressbar/http": "^4.1.1",
-    "@ngx-translate/core": "^9.1.1",
-    "@ngx-translate/http-loader": "^2.0.1",
-    "@swimlane/ngx-charts": "^7.3.0",
-    "angular-highcharts": "^6.0.0",
+    "@ngrx/effects": "7.1.0",
+    "@ngrx/store": "7.1.0",
+    "@ngrx/store-devtools": "7.1.0",
+    "@ngx-progressbar/core": "5.3.2",
+    "@ngx-progressbar/http": "5.3.2",
+    "@ngx-translate/core": "11.0.1",
+    "@ngx-translate/http-loader": "4.0.0",
+    "@swimlane/ngx-charts": "10.0.0",
+    "@types/moment": "^2.13.0",
     "angular2-uuid": "^1.1.1",
     "core-js": "^2.6.4",
-    "countup.js-angular2": "^1.1.1",
-    "d3": "^5.9.0",
+    "countup.js-angular2": "6.1.1",
+    "d3": "5.9.1",
     "hammerjs": "^2.0.8",
     "highcharts": "^6.1.0",
     "lodash": "^4.17.11",
+    "moment": "^2.24.0",
     "ng2-tooltip-directive": "^2.1.6",
-    "ngrx-store-freeze": "^0.2.1",
-    "ngx-moment": "^2.0.0",
+    "ngrx-store-freeze": "0.2.4",
+    "ngx-moment": "3.3.0",
     "npm": "^6.7.0",
-    "rxjs": "^5.5.12",
-    "zone.js": "^0.8.19"
+    "rxjs": "^6.4.0",
+    "tslib": "^1.9.0",
+    "zone.js": "^0.8.29"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/cli": "7.3.1",
+    "@angular/compiler-cli": "7.2.4",
+    "@angular/language-service": "7.2.4",
     "@types/highcharts": "^5.0.36",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.6",
@@ -70,6 +73,6 @@
     "protractor": "^5.4.2",
     "ts-node": "~8.0.2",
     "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "typescript": "3.2.4"
   }
 }
