@@ -1,6 +1,5 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
-import { Observable } from 'rxjs/Observable';
+import { Subscription ,  Observable } from 'rxjs';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
 
 import { SettingsFacade } from '../state/facade.settings';
