@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { EmbyUdpBroadcast } from '../models/emby/emby-udp-broadcast';
 import { EmbyLogin } from '../models/emby//emby-login';
