@@ -1,11 +1,10 @@
+import { throwError, Observable, of } from 'rxjs';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { Actions, Effect } from '@ngrx/effects';
+import { Actions, Effect, ofType } from '@ngrx/effects';
 import { Store } from '@ngrx/store';
 import { map, switchMap, catchError, withLatestFrom } from 'rxjs/operators';
-import { of } from 'rxjs/observable/of';
 
-import 'rxjs/add/observable/throw';
+
 
 import { About } from '../models/about';
 import { AboutService } from '../service/about.service';
@@ -31,8 +30,8 @@ export class AboutEffects {
 
   @Effect()
   getConfiguration$ = this.actions$
-    .ofType(AboutActionTypes.LOAD_ABOUT)
     .pipe(
+      ofType(AboutActionTypes.LOAD_ABOUT),
       map((data: LoadAboutAction) => data.payload),
       withLatestFrom(this.loaded$),
       switchMap(([_, loaded]) => {
@@ -45,6 +44,6 @@ export class AboutEffects {
           ? new LoadAboutSuccessAction(about)
           : new NoNeedAboutAction();
       }),
-      catchError((err: any, caught: Observable<Object>) => Observable.throw(new EffectError(err)))
-  );
+      catchError((err: any, caught: Observable<Object>) => throwError(new EffectError(err)))
+    );
 }
