@@ -1,5 +1,5 @@
 import { Component, Input, OnDestroy } from '@angular/core';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 import { DomSanitizer } from '@angular/platform-browser';
 import { MoviePoster } from '../../models/movie-poster';
 import { SettingsFacade } from '../../../settings/state/facade.settings';
